hotfix(kong-ngx-build) ensure the 'pcre_version' symbol is preserved as undefined when PCRE is statically linked. (#24)
A temporary hotfix until the issue is addressed in upstream OpenResty (a
patch has already been written).

When using `--with-pcre=...`, NGINX is statically linked against
libpcre.a. Since `pcre_version()` is unused, its symbol is stripped by
the linker. Because lua-resty-core's `resty.core.regex` module needs it,
we here ensure that the symbol is entered as undefined in our final
binary.